// src/interfaces/CountriesApi.ts
export interface IApiCountry {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  gini?: Gini;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
}

interface PostalCode {
  format: string;
  regex?: string;
}

interface CapitalInfo {
  latlng?: number[];
}

interface CoatOfArms {
  png?: string;
  svg?: string;
}

interface Flags {
  png: string;
  svg: string;
}

interface Car {
  signs?: string[];
  side: string;
}

interface Gini {
  '2014'?: number;
  '2019'?: number;
  '2018'?: number;
  '2016'?: number;
  '2015'?: number;
  '2003'?: number;
  '2010'?: number;
  '1999'?: number;
  '2017'?: number;
  '2009'?: number;
  '2013'?: number;
  '2011'?: number;
  '1992'?: number;
  '2012'?: number;
  '2006'?: number;
  '1998'?: number;
  '2004'?: number;
  '2005'?: number;
  '2008'?: number;
}

interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

interface Demonyms {
  eng: Eng;
  fra?: Eng;
}

interface Eng {
  f: string;
  m: string;
}

interface Translations {
  ara: Ara;
  bre: Ara;
  ces: Ara;
  cym: Ara;
  deu: Ara;
  est: Ara;
  fin: Ara;
  fra: Ara;
  hrv?: Ara;
  hun: Ara;
  ita: Ara;
  jpn?: Ara;
  kor: Ara;
  nld: Ara;
  per?: Ara;
  pol: Ara;
  por: Ara;
  rus: Ara;
  slk: Ara;
  spa: Ara;
  swe: Ara;
  tur: Ara;
  urd: Ara;
  zho?: Ara;
}

interface Languages {
  ara?: string;
  nld?: string;
  pap?: string;
  grn?: string;
  spa?: string;
  swe?: string;
  div?: string;
  eng?: string;
  mri?: string;
  nzs?: string;
  fra?: string;
  cnr?: string;
  urd?: string;
  crs?: string;
  msa?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  rus?: string;
  uzb?: string;
  bis?: string;
  zho?: string;
  tam?: string;
  ber?: string;
  mey?: string;
  gle?: string;
  kal?: string;
  mfe?: string;
  kaz?: string;
  sqi?: string;
  hmo?: string;
  tpi?: string;
  run?: string;
  hin?: string;
  lav?: string;
  est?: string;
  niu?: string;
  por?: string;
  ell?: string;
  tur?: string;
  glv?: string;
  lit?: string;
  kir?: string;
  mlg?: string;
  hrv?: string;
  dan?: string;
  fao?: string;
  hat?: string;
  rar?: string;
  tet?: string;
  ssw?: string;
  ukr?: string;
  kor?: string;
  aym?: string;
  que?: string;
  arc?: string;
  ckb?: string;
  ron?: string;
  ita?: string;
  khm?: string;
  mlt?: string;
  sin?: string;
  heb?: string;
  tir?: string;
  som?: string;
  pih?: string;
  zdj?: string;
  swa?: string;
  nau?: string;
  gsw?: string;
  roh?: string;
  tgk?: string;
  nor?: string;
  smo?: string;
  ces?: string;
  slk?: string;
  tsn?: string;
  pov?: string;
  mah?: string;
  jam?: string;
  afr?: string;
  nbl?: string;
  nso?: string;
  sot?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zul?: string;
  aze?: string;
  mkd?: string;
  deu?: string;
  ltz?: string;
  kat?: string;
  gil?: string;
  slv?: string;
  tha?: string;
  ind?: string;
  cal?: string;
  cha?: string;
  fin?: string;
  tuk?: string;
  srp?: string;
  lat?: string;
  nrf?: string;
  tvl?: string;
  fas?: string;
  bul?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  nya?: string;
  sna?: string;
  toi?: string;
  zib?: string;
  pol?: string;
  fil?: string;
  nfr?: string;
  ton?: string;
  prs?: string;
  pus?: string;
  sag?: string;
  hye?: string;
  ben?: string;
  lao?: string;
  nep?: string;
  mya?: string;
  hun?: string;
  pau?: string;
  bos?: string;
  dzo?: string;
  cat?: string;
  kon?: string;
  lin?: string;
  lua?: string;
  tkl?: string;
  vie?: string;
  bjz?: string;
  jpn?: string;
  kin?: string;
  isl?: string;
  bel?: string;
  fij?: string;
  hif?: string;
  bar?: string;
  amh?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  mon?: string;
}

interface Idd {
  root?: string;
  suffixes?: string[];
}

interface Currencies {
  MRU?: MRU;
  AWG?: MRU;
  ARS?: MRU;
  SEK?: MRU;
  MVR?: MRU;
  MXN?: MRU;
  NZD?: MRU;
  USD?: MRU;
  XPF?: MRU;
  EUR?: MRU;
  PKR?: MRU;
  ZMW?: MRU;
  SCR?: MRU;
  MYR?: MRU;
  NOK?: MRU;
  UZS?: MRU;
  VUV?: MRU;
  AUD?: MRU;
  SGD?: MRU;
  SRD?: MRU;
  DZD?: MRU;
  MAD?: MRU;
  CRC?: MRU;
  LYD?: MRU;
  QAR?: MRU;
  DKK?: MRU;
  MUR?: MRU;
  KZT?: MRU;
  ALL?: MRU;
  BHD?: MRU;
  PGK?: MRU;
  BIF?: MRU;
  INR?: MRU;
  UYU?: MRU;
  XCD?: MRU;
  BBD?: MRU;
  MOP?: MRU;
  GBP?: MRU;
  IMP?: MRU;
  SYP?: MRU;
  ANG?: MRU;
  XOF?: MRU;
  KGS?: MRU;
  TTD?: MRU;
  EGP?: MRU;
  ILS?: MRU;
  JOD?: MRU;
  MGA?: MRU;
  HRK?: MRU;
  FOK?: MRU;
  HTG?: MRU;
  CKD?: MRU;
  CUC?: MRU;
  CUP?: MRU;
  TWD?: MRU;
  SZL?: MRU;
  ZAR?: MRU;
  UAH?: MRU;
  BMD?: MRU;
  KRW?: MRU;
  PEN?: MRU;
  IQD?: MRU;
  MDL?: MRU;
  VES?: MRU;
  GYD?: MRU;
  KYD?: MRU;
  KHR?: MRU;
  LKR?: MRU;
  SDG?: SDG;
  ERN?: MRU;
  SOS?: MRU;
  KMF?: MRU;
  NIO?: MRU;
  RUB?: MRU;
  UGX?: MRU;
  CHF?: MRU;
  TJS?: MRU;
  SSP?: MRU;
  CZK?: MRU;
  BWP?: MRU;
  TND?: MRU;
  JMD?: MRU;
  MZN?: MRU;
  HNL?: MRU;
  AZN?: MRU;
  MKD?: MRU;
  GMD?: MRU;
  LRD?: MRU;
  CVE?: MRU;
  BSD?: MRU;
  GEL?: MRU;
  KID?: MRU;
  TZS?: MRU;
  CNY?: MRU;
  GTQ?: MRU;
  STN?: MRU;
  SLL?: MRU;
  XAF?: MRU;
  THB?: MRU;
  IDR?: MRU;
  GIP?: MRU;
  TMT?: MRU;
  DOP?: MRU;
  GHS?: MRU;
  JEP?: MRU;
  TRY?: MRU;
  TVD?: MRU;
  IRR?: MRU;
  KES?: MRU;
  BGN?: MRU;
  ZWL?: MRU;
  AOA?: MRU;
  NGN?: MRU;
  PLN?: MRU;
  SHP?: MRU;
  LBP?: MRU;
  WST?: MRU;
  PHP?: MRU;
  GGP?: MRU;
  KWD?: MRU;
  TOP?: MRU;
  OMR?: MRU;
  AFN?: MRU;
  RSD?: MRU;
  AMD?: MRU;
  BDT?: MRU;
  LAK?: MRU;
  CLP?: MRU;
  PAB?: MRU;
  NPR?: MRU;
  MMK?: MRU;
  FKP?: MRU;
  HUF?: MRU;
  GNF?: MRU;
  AED?: MRU;
  BAM?: SDG;
  MWK?: MRU;
  BTN?: MRU;
  CDF?: MRU;
  VND?: MRU;
  BRL?: MRU;
  BZD?: MRU;
  PYG?: MRU;
  BND?: MRU;
  BOB?: MRU;
  DJF?: MRU;
  CAD?: MRU;
  JPY?: MRU;
  RWF?: MRU;
  ISK?: MRU;
  BYN?: MRU;
  HKD?: MRU;
  LSL?: MRU;
  FJD?: MRU;
  COP?: MRU;
  SAR?: MRU;
  ETB?: MRU;
  KPW?: MRU;
  NAD?: MRU;
  MNT?: MRU;
  SBD?: MRU;
  YER?: MRU;
  RON?: MRU;
}

interface SDG {
  name: string;
}

interface MRU {
  name: string;
  symbol: string;
}

interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

interface NativeName {
  ara?: Ara;
  nld?: Ara;
  pap?: Ara;
  grn?: Ara;
  spa?: Ara;
  swe?: Ara;
  div?: Ara;
  eng?: Ara;
  mri?: Ara;
  nzs?: Ara;
  fra?: Ara;
  cnr?: Ara;
  urd?: Ara;
  crs?: Ara;
  msa?: Ara;
  nno?: Ara;
  nob?: Ara;
  smi?: Ara;
  rus?: Ara;
  uzb?: Ara;
  bis?: Ara;
  zho?: Ara;
  tam?: Ara;
  ber?: Ara;
  mey?: Ara;
  gle?: Ara;
  kal?: Ara;
  mfe?: Ara;
  kaz?: Ara;
  sqi?: Ara;
  hmo?: Ara;
  tpi?: Ara;
  run?: Ara;
  hin?: Ara;
  lav?: Ara;
  est?: Ara;
  niu?: Ara;
  por?: Ara;
  ell?: Ara;
  tur?: Ara;
  glv?: Ara;
  lit?: Ara;
  kir?: Ara;
  mlg?: Ara;
  hrv?: Ara;
  dan?: Ara;
  fao?: Ara;
  hat?: Ara;
  rar?: Ara;
  tet?: Ara;
  ssw?: Ara;
  ukr?: Ara;
  kor?: Ara;
  aym?: Ara;
  que?: Ara;
  arc?: Ara;
  ckb?: Ara;
  ron?: Ara;
  ita?: Ara;
  khm?: Ara;
  mlt?: Ara;
  sin?: Ara;
  heb?: Ara;
  tir?: Ara;
  som?: Ara;
  pih?: Ara;
  zdj?: Ara;
  swa?: Ara;
  nau?: Ara;
  gsw?: Ara;
  roh?: Ara;
  tgk?: Ara;
  nor?: Ara;
  smo?: Ara;
  ces?: Ara;
  slk?: Ara;
  tsn?: Ara;
  pov?: Ara;
  mah?: Ara;
  jam?: Ara;
  afr?: Ara;
  nbl?: Ara;
  nso?: Ara;
  sot?: Ara;
  tso?: Ara;
  ven?: Ara;
  xho?: Ara;
  zul?: Ara;
  aze?: Ara;
  mkd?: Ara;
  deu?: Ara;
  ltz?: Ara;
  kat?: Ara;
  gil?: Ara;
  slv?: Ara;
  tha?: Ara;
  ind?: Ara;
  cal?: Ara;
  cha?: Ara;
  fin?: Ara;
  tuk?: Ara;
  srp?: Ara;
  lat?: Ara;
  nrf?: Ara;
  tvl?: Ara;
  fas?: Ara;
  bul?: Ara;
  bwg?: Ara;
  kck?: Ara;
  khi?: Ara;
  ndc?: Ara;
  nde?: Ara;
  nya?: Ara;
  sna?: Ara;
  toi?: Ara;
  zib?: Ara;
  pol?: Ara;
  fil?: Ara;
  nfr?: Ara;
  ton?: Ara;
  prs?: Ara;
  pus?: Ara;
  sag?: Ara;
  hye?: Ara;
  ben?: Ara;
  lao?: Ara;
  nep?: Ara;
  mya?: Ara;
  hun?: Ara;
  pau?: Ara;
  bos?: Ara;
  dzo?: Ara;
  cat?: Ara;
  kon?: Ara;
  lin?: Ara;
  lua?: Ara;
  tkl?: Ara;
  vie?: Ara;
  bjz?: Ara;
  jpn?: Ara;
  kin?: Ara;
  isl?: Ara;
  bel?: Ara;
  fij?: Ara;
  hif?: Ara;
  bar?: Ara;
  amh?: Ara;
  her?: Ara;
  hgm?: Ara;
  kwn?: Ara;
  loz?: Ara;
  ndo?: Ara;
  mon?: Ara;
}

interface Ara {
  official: string;
  common: string;
}